[Feat] : 게시물 삭제 TransactionalEventListener 이용하는 방식으로 수정

+++ src/main/java/com/daangn/dangunmarket/domain/post/service/PostService.java
@@ -121,9 +121,13 @@ public class PostService {
     }
 
     @Transactional
-    public void deletePost(Long postId) {
+    public void deletePost(Long memberId, Long postId) {
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new EntityNotFoundException(NOT_FOUND_POST_ENTITY));
+
+        if (post.isNotOwner(memberId)) {
+            throw new UnauthorizedAccessException(POST_NOT_CREATED_BY_USER);
+        }
 
         post.deletePost();
     }
